[refactor]Spaディレクトリ削除してファイル移動
