+++ app/javascript/packs/router.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import * as VueRouter from 'vue-router';
 import { createRouter, createWebHistory } from 'vue-router'
 
-import About from './components/Spa/About.vue';
+import About from './components/About.vue';
 
 const routes = [
   {
